+++ src/common/file.ts
@@ -1,6 +1,6 @@
 import { exists, extname, resolve } from "trailmix/deps.ts";
 
-import type { FileExtension, ImportOptions } from "trailmix/common/file.d.ts";
+import type { FileExtension, ImportOptions } from "trailmix/common/mod.ts";
 
 const files: Record<FileExtension, string[]> = {
   ts: ["ts", "tsx"],
